+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/transport/TransportCreateRuleAction.java
@@ -38,6 +38,7 @@ import com.wazuh.contentmanager.cti.catalog.service.SecurityAnalyticsException;
 import com.wazuh.contentmanager.cti.catalog.service.SecurityAnalyticsService;
 import com.wazuh.contentmanager.engine.service.EngineService;
 import com.wazuh.contentmanager.rest.model.RestResponse;
+import com.wazuh.contentmanager.settings.PluginSettings;
 import com.wazuh.contentmanager.utils.Constants;
 
 /** Transport action for creating Rule resources. */
@@ -116,6 +117,21 @@ public class TransportCreateRuleAction extends AbstractTransportCreateAction {
         }
 
         return null;
+    }
+
+    @Override
+    protected int getMaxAllowed() {
+        return PluginSettings.getInstance().getMaxRules();
+    }
+
+    @Override
+    protected String getTooManyResourcesMessageFormat() {
+        return Constants.E_400_TOO_MANY_RULES;
+    }
+
+    @Override
+    protected String getMaxReachedLogFormat() {
+        return Constants.I_LOG_MAX_RULES_REACHED;
     }
 
     @Override