+++ CHANGELOG.md
@@ -108,6 +108,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Implement registration based content [(#1138)](https://github.com/wazuh/wazuh-indexer-plugins/pull/1138)
 - Add Data retention policies for stream indices [(#1217)](https://github.com/wazuh/wazuh-indexer-plugins/pull/1217)
 - Add alerting documentation [(#1231)](https://github.com/wazuh/wazuh-indexer-plugins/pull/1231)
+- Add limits for the creation of filters, integrations, decoders, rules and kvdbs [(#1291)](https://github.com/wazuh/wazuh-indexer-plugins/pull/1291)
 - Implement RBAC to safeguard sensitive configuration [(#1310)](https://github.com/wazuh/wazuh-indexer-plugins/pull/1310)
 - Add `index.gc_deletes` setting to stateful index templates [(#1330)](https://github.com/wazuh/wazuh-indexer-plugins/pull/1330)
 
@@ -201,6 +202,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Remove ECS object from WCS definitions [(#612)](https://github.com/wazuh/wazuh-indexer-plugins/pull/612)
 - Remove alerts and archives index creation [(#693)](https://github.com/wazuh/wazuh-indexer-plugins/pull/693)
 - Remove `unclassified` module from WCS and index templates [(#1358)](https://github.com/wazuh/wazuh-indexer-plugins/pull/1358)
+- Remove auto expand replicas [(#1371)](https://github.com/wazuh/wazuh-indexer-plugins/pull/1371)
 
 ### Fixed
 - Improve ECS folder structure [(#473)](https://github.com/wazuh/wazuh-indexer-plugins/pull/473)